refactor: extract application form sections into shared module
Move form state, contact rows and all field sections out of
application-modal.tsx into components/application-form/ so the create
modal and the upcoming detail page share the same building blocks.
Restyle the extracted sections (triage, contacts, documents, resume)
from gray/blue to the nexus slate/indigo tokens and give the notes
and job-description textareas auto-growing, resizable fields.
